Convert mesa host builds to Android.bp
As this is the first conversion in this project, it sets up a new soong
namespace (to be equivalent to the logic around
BOARD_USE_CUSTOMIZED_MESA / BOARD_GPU_DRIVERS).

It also sets up a cc_defaults equivalent to Android.common.mk, and an
aosp-gen-bp.sh script that autogenerates some Android.bp files to be
used when mesa is updated.

Bug: 122332854
Bug: 130696912
Bug: 130723227
Test: lunch aosp_cf_x86_phone-eng; mmm external/mesa3d
Test: treehugger
Change-Id: Ibe5a751e20be1b37be612c9940b49e3c8a88a900